Refact all trackers mocks
