+++ Tests/PlayerTests/Analytics/PlayerItemTrackerTests.swift
@@ -11,20 +11,22 @@ import Foundation
 
 final class PlayerItemTrackerTests: TestCase {
     func testPlayerItemLifecycle() {
-        expectAtLeastEqualPublished(values: [.initialized, .deinitialized], from: SimpleTrackerMock.state) {
+        let publisher = NoMetadataTrackerMock.StatePublisher()
+        expectAtLeastEqualPublished(values: [.initialized, .deinitialized], from: publisher) {
             _ = PlayerItem.simple(
                 url: Stream.shortOnDemand.url,
-                trackerAdapters: [SimpleTrackerMock.adapter()]
+                trackerAdapters: [NoMetadataTrackerMock.adapter(statePublisher: publisher)]
             )
         }
     }
 
     func testItemPlayback() {
         let player = Player()
-        expectEqualPublished(values: [.initialized, .enabled], from: SimpleTrackerMock.state, during: .seconds(2)) {
+        let publisher = NoMetadataTrackerMock.StatePublisher()
+        expectEqualPublished(values: [.initialized, .enabled], from: publisher, during: .seconds(2)) {
             player.append(.simple(
                 url: Stream.onDemand.url,
-                trackerAdapters: [SimpleTrackerMock.adapter()]
+                trackerAdapters: [NoMetadataTrackerMock.adapter(statePublisher: publisher)]
             ))
             player.play()
         }
@@ -32,10 +34,11 @@ final class PlayerItemTrackerTests: TestCase {
 
     func testItemEntirePlayback() {
         let player = Player()
-        expectAtLeastEqualPublished(values: [.initialized, .enabled, .disabled], from: SimpleTrackerMock.state) {
+        let publisher = NoMetadataTrackerMock.StatePublisher()
+        expectAtLeastEqualPublished(values: [.initialized, .enabled, .disabled], from: publisher) {
             player.append(.simple(
                 url: Stream.shortOnDemand.url,
-                trackerAdapters: [SimpleTrackerMock.adapter()]
+                trackerAdapters: [NoMetadataTrackerMock.adapter(statePublisher: publisher)]
             ))
             player.play()
         }
@@ -43,11 +46,12 @@ final class PlayerItemTrackerTests: TestCase {
 
     func testNetworkLoadedItemEntirePlayback() {
         let player = Player()
-        expectAtLeastEqualPublished(values: [.initialized, .enabled, .disabled], from: SimpleTrackerMock.state) {
+        let publisher = NoMetadataTrackerMock.StatePublisher()
+        expectAtLeastEqualPublished(values: [.initialized, .enabled, .disabled], from: publisher) {
             player.append(.mock(
                 url: Stream.shortOnDemand.url,
                 loadedAfter: 1,
-                trackerAdapters: [SimpleTrackerMock.adapter()]
+                trackerAdapters: [NoMetadataTrackerMock.adapter(statePublisher: publisher)]
             ))
             player.play()
         }
@@ -55,10 +59,11 @@ final class PlayerItemTrackerTests: TestCase {
 
     func testFailedItem() {
         let player = Player()
-        expectAtLeastEqualPublished(values: [.initialized, .enabled, .disabled], from: SimpleTrackerMock.state) {
+        let publisher = NoMetadataTrackerMock.StatePublisher()
+        expectAtLeastEqualPublished(values: [.initialized, .enabled, .disabled], from: publisher) {
             player.append(.simple(
                 url: Stream.unavailable.url,
-                trackerAdapters: [SimpleTrackerMock.adapter()]
+                trackerAdapters: [NoMetadataTrackerMock.adapter(statePublisher: publisher)]
             ))
             player.play()
         }
@@ -66,11 +71,12 @@ final class PlayerItemTrackerTests: TestCase {
 
     func testMetadata() {
         let player = Player()
-        expectAtLeastEqualPublished(values: [.initialized, .enabled, .updated("title"), .disabled], from: NonConfigurableTrackerMock.state) {
+        let publisher = TrackerMock<String>.StatePublisher()
+        expectAtLeastEqualPublished(values: [.initialized, .enabled, .updated("title"), .disabled], from: publisher) {
             player.append(.simple(
                 url: Stream.shortOnDemand.url,
                 metadata: AssetMetadataMock(title: "title"),
-                trackerAdapters: [NonConfigurableTrackerMock.adapter { $0.title }]
+                trackerAdapters: [TrackerMock.adapter(statePublisher: publisher) { $0.title }]
             ))
             player.play()
         }
@@ -78,35 +84,14 @@ final class PlayerItemTrackerTests: TestCase {
 
     func testMetadataUpdate() {
         let player = Player()
+        let publisher = TrackerMock<String>.StatePublisher()
         expectAtLeastEqualPublished(
             values: [.initialized, .enabled, .updated("title0"), .updated("title1"), .disabled],
-            from: NonConfigurableTrackerMock.state
+            from: publisher
         ) {
             player.append(.mock(url: Stream.shortOnDemand.url, withMetadataUpdateAfter: 1, trackerAdapters: [
-                NonConfigurableTrackerMock.adapter { $0.title }
+                TrackerMock.adapter(statePublisher: publisher) { $0.title }
             ]))
-            player.play()
-        }
-    }
-
-    func testConfiguration() {
-        let player = Player()
-        expectAtLeastEqualPublished(values: [.initialized("config"), .enabled, .disabled], from: MetadataFreeMock.state) {
-            player.append(.mock(url: Stream.shortOnDemand.url, withMetadataUpdateAfter: 1, trackerAdapters: [
-                MetadataFreeMock.adapter(configuration: "config")
-            ]))
-            player.play()
-        }
-    }
-
-    func testConfigurationAndMetadataUpdate() {
-        let player = Player()
-        expectAtLeastEqualPublished(values: [.initialized("config"), .enabled, .updated("title"), .disabled], from: TrackerMock.state) {
-            player.append(.simple(
-                url: Stream.shortOnDemand.url,
-                metadata: AssetMetadataMock(title: "title"),
-                trackerAdapters: [TrackerMock.adapter(configuration: "config") { $0.title }]
-            ))
             player.play()
         }
     }
