+++ Tests/PlayerTests/Tools/NoMetadataTrackerMock.swift
@@ -8,14 +8,14 @@
 
 import Combine
 
-final class TrackerMock<Metadata: Equatable>: PlayerItemTracker {
+final class NoMetadataTrackerMock: PlayerItemTracker {
     typealias StatePublisher = PassthroughSubject<State, Never>
 
-    enum State: Equatable {
+    enum State {
         case initialized
         case enabled
         case disabled
-        case updated(Metadata)
+        case updated
         case deinitialized
     }
 
@@ -26,11 +26,11 @@ final class TrackerMock<Metadata: Equatable>: PlayerItemTracker {
     private let configuration: Configuration
     private var cancellables = Set<AnyCancellable>()
 
-    init(configuration: Configuration, metadataPublisher: AnyPublisher<Metadata, Never>) {
+    init(configuration: Configuration, metadataPublisher: AnyPublisher<Void, Never>) {
         self.configuration = configuration
         configuration.statePublisher.send(.initialized)
-        metadataPublisher.sink { metadata in
-            configuration.statePublisher.send(.updated(metadata))
+        metadataPublisher.sink { _ in
+            configuration.statePublisher.send(.updated)
         }
         .store(in: &cancellables)
     }
@@ -48,8 +48,8 @@ final class TrackerMock<Metadata: Equatable>: PlayerItemTracker {
     }
 }
 
-extension TrackerMock {
-    static func adapter<M>(statePublisher: StatePublisher, mapper: @escaping (M) -> Metadata) -> TrackerAdapter<M> where M: AssetMetadata {
-        adapter(configuration: Configuration(statePublisher: statePublisher), mapper: mapper)
+extension NoMetadataTrackerMock {
+    static func adapter(statePublisher: StatePublisher) -> TrackerAdapter<Never> {
+        adapter(configuration: Configuration(statePublisher: statePublisher))
     }
 }
